refactor import paths and consolidate configuration
