+++ app/[...puckPath]/client.tsx
@@ -2,7 +2,7 @@
 
 import type { Data } from "@measured/puck";
 import { Render } from "@measured/puck";
-import config from "../../puck.config";
+import config from "@config/page.config";
 
 export function Client({ data }: { data: Data }) {
   return <Render config={config} data={data} />;
